Fix 32 bit compatibility issues (#155)
* Fix obvious printf format incompatibilities reported by the compiler

* config/init: fix invalid value for text_len param in g_scanner_input_text()

The documentation at https://docs.gtk.org/glib/method.Scanner.input_text.html
for g_scanner_input_text() indicates that parameter 2 is a pointer to the
text buffer to be parsed, and parameter 3 is the length of the text buffer.
There is no mention of special values that request the scanner to assume
any special property of the text buffer, such as being a NUL-delimited
string. Therefore, any value other than the true length of the text buffer
risks invoking undefined behavior.

The text_len parameter (parameter 3) of this function is declared to be
a "guint", which is typedef-ed glib/gtypes.h to be a plain "unsigned int".
In both 32-bit and 64-bit gcc, "unsigned int" happens to be a 32-bit value,
as opposed to the compiler pointer types, which are either 32-bit or 64-bit
as required by the target environment.

Internally the code in g_scanner_input_text(), as shown in
https://gitlab.gnome.org/GNOME/glib/-/blob/2.78.4/glib/gscanner.c?ref_type=tags#L1162
uses the text_len to initialize an internal fence pointer to know when to
stop scanning. This fence is calculated by adding text_len to the start of the
buffer. The scanning process is expected to advance the internal start pointer
and terminates when the internal text pointer becomes greater than the
fence pointer calculated at the start.

Prior to this commit, the code at config/init.c called the function
g_scanner_input_text() with a text_len parameter of -1. Inside the function,
this results in a value of 0xFFFFFFFF for the text_len parameter. In a
64-bit environment, this invalid value sort-of-works because it gets added
to a 64-bit pointer value and therefore gets zero-extended to a
fence pointer that starts as greater than the stored start pointer.
However, in a 32-bit environment, the value of 0xFFFFFFFF gets added to
a 32-bit pointer. No zero-extension occurs, and the fence pointer value
instead wraps around to point *before* the stored start pointer. This
results in the end-of-scanning condition getting immediately triggered
and the GLib scanner fails to scan the text buffer. The user-visible symptom
is a failure to recognize any of the data in the config file, just as if
the file was empty, followed by a SIGTRAP.

This commit fixes this by using strlen() on the text buffer instead of -1 so
the text buffer length is always valid.

* Promote input to %llu o long long int

* promote %lld input to long long int

* use %zu for gsize

* promote input to %lld to long long int

* %llu input should be long long unsigned int

* fix paranthethis

* use long long int as input to %lld

---------

+++ modules/mpd.c
@@ -172,9 +172,9 @@ void *mpd_expr_func ( void **params, void *widget, void *event )
   else if(!g_ascii_strcasecmp(params[0],"queue_pos"))
     return g_strdup_printf("%d",status?mpd_status_get_song_pos(status):0);
   else if(!g_ascii_strcasecmp(params[0],"elapsed"))
-    return g_strdup_printf("%lu",status?mpd_status_get_elapsed_ms(status) +
-        (mpd_status_get_state(status)==MPD_STATE_PLAY?
-         (g_get_monotonic_time()-last_update)/1000:0):0);
+    return g_strdup_printf("%llu", (long long unsigned int)(status?
+      (mpd_status_get_elapsed_ms(status) + mpd_status_get_state(status)==MPD_STATE_PLAY?
+         (g_get_monotonic_time()-last_update)/1000:0):0));
   else if(!g_ascii_strcasecmp(params[0],"length"))
     return g_strdup_printf("%u",status?mpd_status_get_total_time(status):0);
   else if(!g_ascii_strcasecmp(params[0],"rate"))
